Fix bug where entering play mode would invalidate the entire mesh cache (with Meshes Are Assets enabled)

+++ probuilder2.0/Assets/ProCore/ProBuilder/Editor/EditorCore/pb_DestroyListener.cs
@@ -40,6 +40,11 @@ namespace ProBuilder2.EditorCommon
 					// otherwise go ahead and destroy as usual
 					if( pb_EditorMeshUtility.GetCachedMesh(pb, out cache_path, out cache_mesh) )
 					{
+						// on entering play mode unity instances everything and destroys the scene,
+						// which nukes the mesh cache.  don't do this.
+						if( EditorApplication.isPlayingOrWillChangePlaymode )
+							return;
+
 						pb_SelectionUtility.Remove(pb);
 						AssetDatabase.DeleteAsset(cache_path);
 					}
